Updated Code to have better measurement
made the measurements more accurate, need to recallibrate camera to have more accuracy but for the moment works well. added in a zfromStart_cm function that shows the distance that the pattern has moved from its original position in the z axis. This currently does not work and needs some editing.